설비 정보 엔티티 생성 / Product 엔티티와 properties create로 수정

+++ src/main/java/com/B1team/b01/entity/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String name;    //제품명
 
     @Column(name="product_price", nullable = false, columnDefinition = "number(10)")
-    private int price;    //가격
+    private Long price;    //가격
 
     @Column(name="product_sort", nullable = false, columnDefinition = "varchar2(15)")
     private String sort;    //분류
